+++ MenuShell/MenuShell.csproj
@@ -47,6 +47,7 @@
     <Compile Include="Domain\Customer.cs" />
     <Compile Include="Domain\LoadCustomers.cs" />
     <Compile Include="Domain\LoadUsers.cs" />
+    <Compile Include="Domain\DeleateUser.cs" />
     <Compile Include="Domain\User.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
